+++ d-mart/src/main/java/com/dmart/model/Store.java
@@ -32,6 +32,6 @@ public class Store {
 	private Address address;
 	
 	@ElementCollection
-	private HashMap<Product,Integer> productQuantity = new HashMap<>();
+	private HashMap<Product,Long> productQuantity = new HashMap<>();
 
 }
